Port to GTK3
The activity was not ported to gtk3.It also contained .pyo files
commited to the repo and .pyo files were also not included in the
.gitignore file.

Porting the activity to gtk3, deleting .pyo files and updating
.gitignore .

Fixes https://github.com/sugarlabs/compress-activity/issues/1
